+++ text-format.lisp
@@ -11,26 +11,26 @@
 
 (defun print-text-format (object &key
                                  (stream *standard-output*)
-                                 (pretty-print t))
+                                 (pretty-print-p t))
     "Prints a protocol buffer message to a stream.
 Parameters:
   OBJECT: The protocol buffer message to print.
   STREAM: The stream to print to.
-  PRETTY-PRINT: When true, generate line breaks and other human readable output
+  PRETTY-PRINT-P: When true, generate line breaks and other human readable output
     in the text format. When false, replace line breaks with spaces."
   (print-text-format-impl object :stream stream
-                                 :pretty-print pretty-print))
+                                 :pretty-print-p pretty-print-p))
 
 (defun print-text-format-impl (object &key
                                       (indent 0)
                                       (stream *standard-output*)
-                                      (pretty-print t))
+                                      (pretty-print-p t))
   "Prints a protocol buffer message to a stream.
 Parameters:
   OBJECT: The protocol buffer message to print.
   INDENT: Indent the output by INDENT spaces. Only used for pretty-printing.
   STREAM: The stream to print to.
-  PRETTY-PRINT: When true, generate line breaks and other human readable output
+  PRETTY-PRINT-P: When true, generate line breaks and other human readable output
     in the text format. When false, replace line breaks with spaces."
   (let* ((type (type-of object))
          (message (find-message-descriptor type :error-p t)))
@@ -49,13 +49,13 @@ Parameters:
                                   (proto-name field)
                                   :indent indent
                                   :stream stream
-                                  :pretty-print pretty-print))
+                                  :pretty-print-p pretty-print-p))
               (print-field value
                            (proto-class field)
                            (proto-name field)
                            :indent indent
                            :stream stream
-                           :pretty-print pretty-print)))))
+                           :pretty-print-p pretty-print-p)))))
     (dolist (oneof (proto-oneofs message))
       (let* ((oneof-data (slot-value object (oneof-descriptor-internal-name oneof)))
              (set-field (oneof-set-field oneof-data)))
@@ -66,11 +66,11 @@ Parameters:
                          (proto-name field-desc)
                          :indent indent
                          :stream stream
-                         :pretty-print pretty-print)))))
+                         :pretty-print-p pretty-print-p)))))
     nil))
 
 (defun print-field (value type name
-                    &key (indent 0) (stream *standard-output*) (pretty-print t))
+                    &key (indent 0) (stream *standard-output*) (pretty-print-p t))
   "Print the text format of a single field which is not repeated.
 Parameters:
   VALUE: The value in the field to print.
@@ -80,7 +80,7 @@ Parameters:
   INDENT: If supplied, indent the text by INDENT spaces.
   STREAM: The stream to output to.
   PRINT-NAME: Whether or not to print the name of the field.
-  PRETTY-PRINT: When true, print newlines and indentation."
+  PRETTY-PRINT-P: When true, print newlines and indentation."
   ;; If VALUE is NIL and the type is not boolean, there is nothing to do.
   (unless (or value (eq type 'boolean) (eq type 'symbol))
     (return-from print-field nil))
@@ -88,29 +88,29 @@ Parameters:
     (cond
       ((scalarp type)
        (print-scalar value type name stream
-                     (and pretty-print indent)))
+                     (and pretty-print-p indent)))
       ((typep (setq desc (or (find-message-descriptor type)
                              (find-enum-descriptor type)
                              (find-map-descriptor type)))
               'message-descriptor)
-       (print-message-brace t name pretty-print indent stream)
+       (print-message-brace t name pretty-print-p indent stream)
        (print-text-format-impl value :indent (+ indent 2)
                                      :stream stream
-                                     :pretty-print pretty-print)
-       (print-message-brace nil name pretty-print indent stream))
+                                     :pretty-print-p pretty-print-p)
+       (print-message-brace nil name pretty-print-p indent stream))
       ((typep desc 'enum-descriptor)
-       (print-enum value desc name stream (and pretty-print indent)))
+       (print-enum value desc name stream (and pretty-print-p indent)))
       ((typep desc 'map-descriptor)
        (loop for k being the hash-keys of value using (hash-value v)
-             do (if pretty-print
+             do (if pretty-print-p
                     (format stream "~&~V,0T~A { " indent name)
                     (format stream "~A { " name))
                 (print-scalar k (proto-key-type desc) "key" stream nil)
                 (print-field v (proto-value-type desc) "value"
                              :stream stream
-                             :pretty-print nil)
+                             :pretty-print-p nil)
                 (format stream "}")
-                (when pretty-print
+                (when pretty-print-p
                   (format stream "~%"))))
       ;; This case only happens when the user specifies a custom type and
       ;; doesn't support it above.
@@ -188,20 +188,20 @@ Parameters:
           (format stream "~%")
           (format stream " ")))))
 
-(defun print-message-brace (opening-p name pretty-print indent stream)
+(defun print-message-brace (opening-p name pretty-print-p indent stream)
   "Print either the opening NAME { or closing }.
 
 Parameters:
   OPENING-P: Is this an opening or closing brace.
   NAME: The name to print before the value. If NIL, no name will be printed.
-  PRETTY-PRINT: When true, print newlines and indentation.
+  PRETTY-PRINT-P: When true, print newlines and indentation.
   INDENT: A set indentation to print to. Used only for pretty-print.
   STREAM: The stream to print to."
   (if opening-p
-      (if pretty-print
+      (if pretty-print-p
           (format stream "~&~V,0T~A {~%" indent name)
           (format stream "~A { " name))
-      (if pretty-print
+      (if pretty-print-p
           (format stream "~&~V,0T}~%" indent)
           (format stream "} "))))
 
@@ -347,4 +347,4 @@ return T as a second value."
   (declare (ignore width))
   (cond (other-args (error "FORMAT directive ~~/cl-protobufs:fmt/ takes only one argument."))
         (colon-p (error "FORMAT directive ~~/cl-protobufs:fmt/ does not take colons."))
-        (t (print-text-format proto :stream stream :pretty-print at-sign-p))))
+        (t (print-text-format proto :stream stream :pretty-print-p at-sign-p))))